requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
gensim
nltk
wordcloud
python-dotenv
praw
tweepy
